Fixed warning for Win 64 bits compilation. Clarified Verbose levels. Disabled balancing if run serial. Move various initialization to be constructor initialization lists. Fixed storage of results in MPI run. All messages now go to cerr. Different initialization of proportions. Now all classes have the same value.
git-svn-id: https://svn.vital-it.ch/svn/hp2c/trunk/Codeml_Variants/Fastcodeml@3700 95c5a404-1f88-0410-a6b0-c3f062b6f34a

+++ Forest.h
@@ -190,7 +190,7 @@ public:
 	///
 	/// @return The node index corresponding to the foreground branch
 	///
-	unsigned int adjustFgBranchIdx(unsigned int aFgBranch) const {return mTableInternalToBranchID[aFgBranch];}
+	unsigned int adjustFgBranchIdx(size_t aFgBranch) const {return mTableInternalToBranchID[aFgBranch];}
 
 	/// Access the global list of node names.
 	///
